Use uint8 instead of int for turn index slicing

+++ gojax/go.py
@@ -32,7 +32,7 @@ def at_pieces_per_turn(states, turns):
     :param turns: a boolean array of length N indicating which pieces to reference per state.
     :return: an update reference array of shape N x B x B.
     """
-    return states.at[jnp.arange(states.shape[0]), jnp.array(turns, dtype=int)]
+    return states.at[jnp.arange(states.shape[0]), jnp.array(turns, dtype='uint8')]
 
 
 def at_location_per_turn(states, turns, row, col):
@@ -48,8 +48,8 @@ def at_location_per_turn(states, turns, row, col):
     :return: a scalar update reference.
     """
     return states.at[
-        jnp.arange(states.shape[0]), jnp.array(turns, dtype=int), jnp.full(states.shape[0],
-                                                                           row), jnp.full(
+        jnp.arange(states.shape[0]), jnp.array(turns, dtype='uint8'), jnp.full(states.shape[0],
+                                                                               row), jnp.full(
             states.shape[0], col)]
